Add dynamic lookup
mmdb:LookupField(IP, "1", "2", "3") This function allows you to lookup the same data as the other functions but also allows you to lookup more data. IF the mmdb file has the data.
There was also a thing that I overlooked, an error caused by copilot returning a reference instead of the pointer. This caused it to crash.