update
Remove the new and delete because there are better ways.
Added the close function because we are not managing the memory ourselves.
Added GetIPCoordinates and GetIPCityName which are only available in the City to IP database.
Added a bunch of comments to make it slightly easier to update in the future.